+++ packages/components/src/toggle-group-control/test/index.tsx
@@ -2,7 +2,7 @@
  * External dependencies
  */
 import { render, screen, waitFor } from '@testing-library/react';
-import userEvent from '@testing-library/user-event';
+import { press, click, hover, sleep } from '@ariakit/test';
 
 /**
  * WordPress dependencies
@@ -19,8 +19,13 @@ import {
 	ToggleGroupControlOption,
 	ToggleGroupControlOptionIcon,
 } from '../index';
+import { TOOLTIP_DELAY } from '../../tooltip';
 import type { ToggleGroupControlProps } from '../types';
-import cleanupTooltip from '../../tooltip/test/utils';
+
+const hoverOutside = async () => {
+	await hover( document.body );
+	await hover( document.body, { clientX: 10, clientY: 10 } );
+};
 
 const ControlledToggleGroupControl = ( {
 	value: valueProp,
@@ -113,7 +118,6 @@ describe.each( [
 		} );
 	} );
 	it( 'should call onChange with proper value', async () => {
-		const user = userEvent.setup();
 		const mockOnChange = jest.fn();
 
 		render(
@@ -126,13 +130,12 @@ describe.each( [
 			</Component>
 		);
 
-		await user.click( screen.getByRole( 'radio', { name: 'R' } ) );
+		await click( screen.getByRole( 'radio', { name: 'R' } ) );
 
 		expect( mockOnChange ).toHaveBeenCalledWith( 'rigas' );
 	} );
 
 	it( 'should render tooltip where `showTooltip` === `true`', async () => {
-		const user = userEvent.setup();
 		render(
 			<Component label="Test Toggle Group Control">
 				{ optionsWithTooltip }
@@ -143,19 +146,26 @@ describe.each( [
 			'Click for Delicious Gnocchi'
 		);
 
-		await user.hover( firstRadio );
+		await hover( firstRadio );
 
-		const tooltip = await screen.findByText(
-			'Click for Delicious Gnocchi'
-		);
+		const tooltip = await screen.findByRole( 'tooltip', {
+			name: 'Click for Delicious Gnocchi',
+		} );
 
 		await waitFor( () => expect( tooltip ).toBeVisible() );
 
-		await cleanupTooltip( user );
+		// hover outside of radio
+		await hoverOutside();
+
+		// Tooltip should hide
+		expect(
+			screen.queryByRole( 'tooltip', {
+				name: 'Click for Delicious Gnocchi',
+			} )
+		).not.toBeInTheDocument();
 	} );
 
 	it( 'should not render tooltip', async () => {
-		const user = userEvent.setup();
 		render(
 			<Component label="Test Toggle Group Control">
 				{ optionsWithTooltip }
@@ -166,19 +176,24 @@ describe.each( [
 			'Click for Sumptuous Caponata'
 		);
 
-		await user.hover( secondRadio );
+		await hover( secondRadio );
 
-		await waitFor( () =>
-			expect(
-				screen.queryByText( 'Click for Sumptuous Caponata' )
-			).not.toBeInTheDocument()
-		);
+		// Tooltip shouldn't show
+		expect(
+			screen.queryByText( 'Click for Sumptuous Caponata' )
+		).not.toBeInTheDocument();
+
+		// Advance time by default delay
+		await sleep( TOOLTIP_DELAY );
+
+		// Tooltip shouldn't show.
+		expect(
+			screen.queryByText( 'Click for Sumptuous Caponata' )
+		).not.toBeInTheDocument();
 	} );
 
 	if ( mode === 'controlled' ) {
 		it( 'should reset values correctly', async () => {
-			const user = userEvent.setup();
-
 			render(
 				<Component label="Test Toggle Group Control">
 					{ options }
@@ -188,25 +203,23 @@ describe.each( [
 			const rigasOption = screen.getByRole( 'radio', { name: 'R' } );
 			const jackOption = screen.getByRole( 'radio', { name: 'J' } );
 
-			await user.click( rigasOption );
+			await click( rigasOption );
 
 			expect( jackOption ).not.toBeChecked();
 			expect( rigasOption ).toBeChecked();
 
-			await user.keyboard( '[ArrowRight]' );
+			await press.ArrowRight();
 
 			expect( rigasOption ).not.toBeChecked();
 			expect( jackOption ).toBeChecked();
 
-			await user.click( screen.getByRole( 'button', { name: 'Reset' } ) );
+			await click( screen.getByRole( 'button', { name: 'Reset' } ) );
 
 			expect( rigasOption ).not.toBeChecked();
 			expect( jackOption ).not.toBeChecked();
 		} );
 
 		it( 'should update correctly when triggered by external updates', async () => {
-			const user = userEvent.setup();
-
 			render(
 				<Component
 					value="rigas"
@@ -225,19 +238,19 @@ describe.each( [
 				screen.getByRole( 'radio', { name: 'J' } )
 			).not.toBeChecked();
 
-			await user.click( screen.getByRole( 'button', { name: 'Jack' } ) );
+			await click( screen.getByRole( 'button', { name: 'Jack' } ) );
 			expect( screen.getByRole( 'radio', { name: 'J' } ) ).toBeChecked();
 			expect(
 				screen.getByRole( 'radio', { name: 'R' } )
 			).not.toBeChecked();
 
-			await user.click( screen.getByRole( 'button', { name: 'Rigas' } ) );
+			await click( screen.getByRole( 'button', { name: 'Rigas' } ) );
 			expect( screen.getByRole( 'radio', { name: 'R' } ) ).toBeChecked();
 			expect(
 				screen.getByRole( 'radio', { name: 'J' } )
 			).not.toBeChecked();
 
-			await user.click( screen.getByRole( 'button', { name: 'Reset' } ) );
+			await click( screen.getByRole( 'button', { name: 'Reset' } ) );
 			expect(
 				screen.getByRole( 'radio', { name: 'R' } )
 			).not.toBeChecked();
@@ -251,7 +264,6 @@ describe.each( [
 		describe( 'isDeselectable = false', () => {
 			it( 'should not be deselectable', async () => {
 				const mockOnChange = jest.fn();
-				const user = userEvent.setup();
 
 				render(
 					<Component
@@ -267,31 +279,35 @@ describe.each( [
 					name: 'R',
 					checked: true,
 				} );
-				await user.click( rigas );
+				await click( rigas );
 				expect( mockOnChange ).toHaveBeenCalledTimes( 0 );
 			} );
 
 			it( 'should not tab to next radio option', async () => {
-				const user = userEvent.setup();
-
 				render(
-					<Component value="rigas" label="Test">
-						{ options }
-					</Component>
+					<>
+						<Component value="rigas" label="Test">
+							{ options }
+						</Component>
+						<button>After ToggleGroupControl</button>
+					</>
 				);
 
 				const rigas = screen.getByRole( 'radio', {
 					name: 'R',
 				} );
 
-				await user.tab();
+				await press.Tab();
 				expect( rigas ).toHaveFocus();
 
-				await user.tab();
+				await press.Tab();
 
+				// When in controlled mode, there is an additional "Reset" button.
 				const expectedFocusTarget =
 					mode === 'uncontrolled'
-						? rigas.ownerDocument.body
+						? screen.getByRole( 'button', {
+								name: 'After ToggleGroupControl',
+						  } )
 						: screen.getByRole( 'button', { name: 'Reset' } );
 
 				expect( expectedFocusTarget ).toHaveFocus();
@@ -301,7 +317,6 @@ describe.each( [
 		describe( 'isDeselectable = true', () => {
 			it( 'should be deselectable', async () => {
 				const mockOnChange = jest.fn();
-				const user = userEvent.setup();
 
 				render(
 					<Component
@@ -314,7 +329,7 @@ describe.each( [
 					</Component>
 				);
 
-				await user.click(
+				await click(
 					screen.getByRole( 'button', {
 						name: 'R',
 						pressed: true,
@@ -323,7 +338,7 @@ describe.each( [
 				expect( mockOnChange ).toHaveBeenCalledTimes( 1 );
 				expect( mockOnChange ).toHaveBeenLastCalledWith( undefined );
 
-				await user.click(
+				await click(
 					screen.getByRole( 'button', {
 						name: 'R',
 						pressed: false,
@@ -334,15 +349,13 @@ describe.each( [
 			} );
 
 			it( 'should tab to the next option button', async () => {
-				const user = userEvent.setup();
-
 				render(
 					<Component isDeselectable value="rigas" label="Test">
 						{ options }
 					</Component>
 				);
 
-				await user.tab();
+				await press.Tab();
 				expect(
 					screen.getByRole( 'button', {
 						name: 'R',
@@ -350,7 +363,7 @@ describe.each( [
 					} )
 				).toHaveFocus();
 
-				await user.tab();
+				await press.Tab();
 				expect(
 					screen.getByRole( 'button', {
 						name: 'J',
@@ -359,7 +372,7 @@ describe.each( [
 				).toHaveFocus();
 
 				// Focus should not move with arrow keys
-				await user.keyboard( '{ArrowLeft}' );
+				await press.ArrowLeft();
 				expect(
 					screen.getByRole( 'button', {
 						name: 'J',